third_party/blink/web_tests/platform/generic/fast/css/invalid-hex-color-expected.txt

Test 0 (Invalid: 0 hexadecimal digits) [EXPECTED] [ACTUAL] PASS
Test 1 (Invalid: 1 hexadecimal digits) [EXPECTED] [ACTUAL] PASS
Test 2 (Invalid: 2 hexadecimal digits) [EXPECTED] [ACTUAL] PASS
Test 3 (Valid: 3 hexadecimal digits) [EXPECTED] [ACTUAL] PASS
Test 4 (Valid: 4 hexadecimal digits) [EXPECTED] [ACTUAL] PASS
Test 5 (Invalid: 5 hexadecimal digits) [EXPECTED] [ACTUAL] PASS
Test 6 (Valid: 6 hexadecimal digits) [EXPECTED] [ACTUAL] PASS
Test 7 (Invalid: 7 hexadecimal digits) [EXPECTED] [ACTUAL] PASS
Test 8 (Valid: 8 hexadecimal digits) [EXPECTED] [ACTUAL] PASS
Test 9 (Invalid: 0 hexadecimal digits concatenated with 'px') [EXPECTED] [ACTUAL] PASS
Test 10 (Invalid: 1 hexadecimal digits concatenated with 'px') [EXPECTED] [ACTUAL] PASS
Test 11 (Invalid: 2 hexadecimal digits concatenated with 'px') [EXPECTED] [ACTUAL] PASS
Test 12 (Invalid: 3 hexadecimal digits concatenated with 'px') [EXPECTED] [ACTUAL] PASS
Test 13 (Invalid: 4 hexadecimal digits concatenated with 'px') [EXPECTED] [ACTUAL] PASS
Test 14 (Invalid: 5 hexadecimal digits concatenated with 'px') [EXPECTED] [ACTUAL] PASS
Test 15 (Invalid: 6 hexadecimal digits concatenated with 'px') [EXPECTED] [ACTUAL] PASS
Test 16 (Invalid: 7 hexadecimal digits concatenated with 'px') [EXPECTED] [ACTUAL] PASS
Test 17 (Invalid: 8 hexadecimal digits concatenated with 'px') [EXPECTED] [ACTUAL] PASS
